--- a/test/printf-test.cc
+++ b/test/printf-test.cc
@@ -617,7 +617,9 @@ TEST(PrintfTest, PrintfDetermineOutputSize) {
 
   const auto format_string = "%s";
   const auto format_arg = "Hello";
-  const auto expected_size = fmt::sprintf(format_string, format_arg).size();
+  const auto expected_size = 5;
+
+  EXPECT_EQ(expected_size, fmt::sprintf(format_string, format_arg).size());
 
   EXPECT_EQ(expected_size,
             (truncated_printf_context(
@@ -625,20 +627,4 @@ TEST(PrintfTest, PrintfDetermineOutputSize) {
                  fmt::make_format_args<truncated_printf_context>(format_arg))
                  .format()
                  .count()));
-}
-
-TEST(PrintfTest, PrintfAppendToBuffer) {
-  using backit = std::back_insert_iterator<fmt::basic_memory_buffer<char>>;
-  using context = fmt::basic_printf_context<backit, char>;
-
-  const auto format_string = "%s";
-  const char* format_arg = "Hello";
-  fmt::basic_memory_buffer<char> buffer;
-  context(std::back_inserter(buffer), format_string,
-          fmt::make_format_args<context>(format_arg))
-      .format();
-
-  std::string result(std::begin(buffer), std::end(buffer));
-
-  EXPECT_EQ(std::string("Hello"), result);
 }