<!-- Please read the contribution guidelines before submitting a pull request. -->
<!-- By submitting this pull request, you agree that your contributions are licensed under the {fmt} license,
     and agree to future changes to the licensing. -->
<!-- If you're a first-time contributor, please acknowledge it by leaving the statement below. -->

I agree that my contributions are licensed under the {fmt} license, and agree to future changes to the licensing.

This affects the performance of fmt::sprintf in the case where the format string is at least on the order of 100 characters long.

The benchmark:

```c++
#include <benchmark/benchmark.h>
#include <fmt/printf.h>

void long_printf_format_string(benchmark::State& state) {
  std::string format_string = std::string(state.range(0), 'A') + "%s";
  std::string message = "Hi";
  for (auto _ : state) {
    benchmark::DoNotOptimize(fmt::sprintf(format_string, message));
  };
}
BENCHMARK(long_printf_format_string)->Arg(10)->Arg(100)->Arg(1000)->Arg(10000);

void long_fmt_format_string(benchmark::State& state) {
  std::string format_string = std::string(state.range(0), 'A') + "{}";
  std::string message = "Hi";
  for (auto _ : state) {
    benchmark::DoNotOptimize(fmt::format(format_string, message));
  };
}
BENCHMARK(long_fmt_format_string)->Arg(10)->Arg(100)->Arg(1000)->Arg(10000);
```

The benchmark numbers change as follows:

master (60dc273513449dc959fc8212464058c54c7c36bb)
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
long_printf_format_string/10          39.9 ns         39.9 ns     17364987
long_printf_format_string/100          188 ns          188 ns      3697522
long_printf_format_string/1000        1402 ns         1402 ns       494033
long_printf_format_string/10000      12409 ns        12409 ns        55024
long_fmt_format_string/10             28.0 ns         28.0 ns     24726944
long_fmt_format_string/100            88.2 ns         88.2 ns      7890169
long_fmt_format_string/1000            333 ns          333 ns      2119728
long_fmt_format_string/10000           873 ns          873 ns       779553
```
branch (104b3de77b637f94334dcfd726abe3d38c3d13c)
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
long_printf_format_string/10          36.7 ns         36.7 ns     17982978
long_printf_format_string/100         95.8 ns         95.8 ns      7214487
long_printf_format_string/1000         337 ns          337 ns      2075802
long_printf_format_string/10000        763 ns          763 ns       910155
long_fmt_format_string/10             28.0 ns         28.0 ns     24945565
long_fmt_format_string/100            86.6 ns         86.6 ns      7932910
long_fmt_format_string/1000            330 ns          330 ns      2093427
long_fmt_format_string/10000           838 ns          838 ns       807894
```