@vitaut 2020-11-05
Thanks for the PR. Please fix formatting in the PR description and post benchmark results as text, not as image.