@rimathia 2020-11-05
Thanks for getting to the PR so quickly. Just tell me if you'd like to have the json output as well.