@rimathia 2020-11-08
I'll have to get back to this next week. Just to be sure I'm not operating on wrong assumptions: The following two benchmarks should have very similar results, right?

```c++
void long_printf_format_string(benchmark::State& state) {
  std::string format_string = std::string(state.range(0), 'A') + "%s";
  std::string message = "Hi";
  for (auto _ : state) {
    benchmark::DoNotOptimize(fmt::sprintf(format_string, message));
  };
}
BENCHMARK(long_printf_format_string)->Arg(10)->Arg(100)->Arg(1000)->Arg(10000);

void long_fmt_format_string(benchmark::State& state) {
  std::string format_string = std::string(state.range(0), 'A') + "{}";
  std::string message = "Hi";
  for (auto _ : state) {
    benchmark::DoNotOptimize(fmt::format(format_string, message));
  };
}
BENCHMARK(long_fmt_format_string)->Arg(10)->Arg(100)->Arg(1000)->Arg(10000);
```
On the branch (c4f62bf21158ec006f07cae373436635a492464e) the version with printf becomes quite a bit faster than the fmt version from 100 characters:
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
long_printf_format_string/10          45.7 ns         45.6 ns     14294174
long_printf_format_string/100          116 ns          116 ns      6097242
long_printf_format_string/1000         405 ns          405 ns      1756349
long_printf_format_string/10000        870 ns          869 ns       775744
long_fmt_format_string/10             42.0 ns         42.0 ns     16111176
long_fmt_format_string/100             237 ns          237 ns      2903444
long_fmt_format_string/1000           1696 ns         1695 ns       401220
long_fmt_format_string/10000         13772 ns        13755 ns        50931
```