@vitaut 2020-11-08
> The following two benchmarks should have very similar results, right?

Somewhat similar. The printf numbers look suspicious both compared to format and your earlier results. I recommend checking that the modified version produces the correct result.