@vitaut 2020-11-11
I finally got a chance to look into this and the results are indeed correct. As it turned out `format_handler::on_text` wasn't going through the fast path which I fixed in https://github.com/fmtlib/fmt/commit/beb248b6ac4d366783305a25af47d52b03181cf8:

```
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
long_fmt_format_string           899 ns          899 ns       749016
```

Thanks for catching this!

My earlier comment about using `fmt::detail::write` still holds.